requirements.txt
APScheduler
asgiref
Django
django-apscheduler
Pillow
psycopg2-binary
python-dotenv
pytz
six
sqlparse
tzdata
tzlocal